Add more testing and modificated the toll to improve security and save to wrong behavior

+++ models/tolls.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose');
+
 const Schema = mongoose.Schema;
 
 const TollSchema = new Schema({
@@ -14,14 +15,19 @@ const TollSchema = new Schema({
   group: { type: Number, default: 2 }
 }, { versionKey: false });
 
-// check this function
+
 TollSchema.statics.createToll = async function (toll) {
   if (Object.keys(toll).length === 0) return null;
   return await this.create(toll);
 };
 
 TollSchema.statics.deleteToll = async function (id) {
-  await this.findByIdAndDelete(id);
+  if (typeof(id) !== typeof({})) { return null; }
+  if (mongoose.Types.ObjectId.isValid(id)) {
+    await this.findByIdAndDelete(id)
+    return 'Nan'
+  }
+  return null
 };
 
 TollSchema.statics.findTollById = async function (id, callback) {
@@ -39,6 +45,8 @@ TollSchema.statics.getTolls = async function () {
 
 TollSchema.statics.updateToll = async function (id, data, callback) {
   var toll = null;
+  if (typeof(data) !== typeof({})) return null;
+  if (Object.keys(data).length === 0) return null;
   if (mongoose.isValidObjectId(id)) { toll = await this.findByIdAndUpdate(id, data, { new: true }); }
   return toll;
 };
